+++ src/firebolt.cpp
@@ -18,6 +18,7 @@
 
 #include "firebolt/firebolt.h"
 #include "accessibility_impl.h"
+#include "actions_impl.h"
 #include "advertising_impl.h"
 #include "device_impl.h"
 #include "discovery_impl.h"
@@ -40,6 +41,7 @@ public:
     FireboltAccessorImpl()
         : accessibility_(Firebolt::Helpers::GetHelperInstance()),
           advertising_(Firebolt::Helpers::GetHelperInstance()),
+          actions_(Firebolt::Helpers::GetHelperInstance()),
           device_(Firebolt::Helpers::GetHelperInstance()),
           discovery_(Firebolt::Helpers::GetHelperInstance()),
           display_(Firebolt::Helpers::GetHelperInstance()),
@@ -83,11 +85,13 @@ public:
     Presentation::IPresentation& PresentationInterface() override { return presentation_; }
     Stats::IStats& StatsInterface() override { return stats_; }
     TextToSpeech::ITextToSpeech& TextToSpeechInterface() override { return textToSpeech_; }
+    Actions::IActions& ActionsInterface() override { return actions_; }
 
 private:
     void unsubscribeAll()
     {
         accessibility_.unsubscribeAll();
+        actions_.unsubscribeAll();
         lifecycle_.unsubscribeAll();
         localization_.unsubscribeAll();
         network_.unsubscribeAll();
@@ -98,6 +102,7 @@ private:
 private:
     Accessibility::AccessibilityImpl accessibility_;
     Advertising::AdvertisingImpl advertising_;
+    Actions::ActionsImpl actions_;
     Device::DeviceImpl device_;
     Discovery::DiscoveryImpl discovery_;
     Display::DisplayImpl display_;