+++ test/api_test_app/apis/discoveryDemo.cpp
@@ -61,7 +61,7 @@ void DiscoveryDemo::runOption(const std::string& method)
                                                                                       watchedOn, agePolicyOpt);
         if (succeed(r))
         {
-            std::cout << "Discovery.watched result: " << std::boolalpha << *r << std::endl;
+            std::cout << "Discovery.watched: Success" << std::endl;
         }
     }
 }