+++ test/unit/mock_helper.h
@@ -104,6 +104,28 @@ protected:
                              { return Firebolt::Result<nlohmann::json>{response}; }));
     }
 
+    void mockInvoke(const std::string& methodName)
+    {
+        EXPECT_CALL(mockHelper, invoke(methodName, _))
+            .WillOnce(Invoke(
+                [&](const std::string& invokedMethodName, const nlohmann::json& parameters)
+                {
+                    nlohmann::json message = {
+                        {"jsonrpc", "2.0"},
+                        {"id", "0"},
+                        {"method", invokedMethodName},
+                    };
+
+                    if (!parameters.is_null())
+                    {
+                        message["params"] = parameters;
+                    }
+
+                    Firebolt::Error err = jsonEngine.MockResponse(message);
+                    return Firebolt::Result<void>{err};
+                }));
+    }
+
     void mockSubscribe(const std::string& eventName)
     {
         EXPECT_CALL(mockHelper, subscribe(_, eventName, _, _))